added prev and next, moved select all
