+++ src/components/IntentSelection/IntentList/Intent/Intent.style.jsx
@@ -3,7 +3,7 @@ import { colors } from '../../../../utils/constants';
 const { accent, gray } = colors;
 
 export const getCardStyle = (selected) => ({
-  width: 410,
+  width: 380,
   margin: 20,
   borderRadius: 20,
   textAlign: 'left',
