+++ src/components/IntentSelection/IntentList/IntentList.style.jsx
@@ -1,0 +1,6 @@
+export const selectButton = {
+  marginRight: 12,
+  width: 140,
+};
+
+export default { selectButton };
